Visible numbers of subspaces and devices
