+++ src/app/Devices.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import { SpaceResourceApi, DeviceDTO } from "./srvapi";
-import { withStyles, WithStyles, createStyles, Theme } from '@material-ui/core/styles';
-
-const styles = (theme: Theme) => createStyles({
-});
+import {
+    withStyles, WithStyles, Accordion,
+    AccordionSummary, AccordionDetails, Typography
+} from "@material-ui/core";
+import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import styles from "./Styles";
 
 interface DevicesProps extends WithStyles<typeof styles> {
     spaceId: number;
@@ -40,11 +42,46 @@ class Devices extends React.Component<DevicesProps, DevicesState> {
         return (
             <>
                 {
-                    this.state.devices.map(device => (
-                        <p key={device.id}>
-                            {device.slug}
-                        </p>
-                    ))
+                    this.state.devices.length > 0 ?
+                        <Accordion>
+                            <AccordionSummary
+                                expandIcon={<ExpandMoreIcon />}
+                                aria-controls="panel2bh-content"
+                                id="panel2bh-header"
+                            >
+                                <Typography className={this.props.classes.heading}>Devices</Typography>
+                                <Typography className={this.props.classes.secondaryHeading}>{this.state.devices.length}</Typography>
+                            </AccordionSummary>
+                            <AccordionDetails>
+                                {
+                                    this.state.devices.map(device => (
+                                        <p key={device.id}>
+                                            {device.slug}
+                                        </p>
+                                    ))
+                                }
+                            </AccordionDetails>
+                        </Accordion>
+                        :
+                        <Accordion disabled>
+                            <AccordionSummary
+                                expandIcon={<ExpandMoreIcon />}
+                                aria-controls="panel2bh-content"
+                                id="panel2bh-header"
+                            >
+                                <Typography className={this.props.classes.heading}>Devices</Typography>
+                                <Typography className={this.props.classes.secondaryHeading}>{this.state.devices.length}</Typography>
+                            </AccordionSummary>
+                            <AccordionDetails>
+                                {
+                                    this.state.devices.map(device => (
+                                        <p key={device.id}>
+                                            {device.slug}
+                                        </p>
+                                    ))
+                                }
+                            </AccordionDetails>
+                        </Accordion>
                 }
             </>
         )
